added back in opening cutscene and moved functions from home.js to a generic map.js in preperation for more maps.
