+++ src/maps/home.js
@@ -1,11 +1,11 @@
 import Hero from '../sprites/hero.js';
-import NPC from '../sprites/npc.js'
-import OverworldEvent from '../overworldEvent.js';
+import NPC from '../sprites/npc.js';
+import Map from './map.js'
 
-export default class Home {
+export default class Home extends Map {
   constructor() {
+    super();
     this.id = 'Home';
-    this.isCutScenePlaying = false;
   }
 
   initialize(scene) {
@@ -63,106 +63,37 @@ export default class Home {
      * --- CUTSCENES ---
      */
     this.openingCutScene = [
-      // {who: 'hero', type: 'walk', direction: 'right'},
-      // {who: 'hero', type: 'walk', direction: 'right'},
-      // {who: 'hero', type: 'walk', direction: 'right'},
-      // {who: 'hero', type: 'walk', direction: 'down'},
-      // {who: 'hero', type: 'walk', direction: 'down'},
-      // {who: 'villager', type: 'walk', direction: 'left'},
-      // {who: 'villager', type: 'walk', direction: 'left'},
-      // {who: 'villager', type: 'walk', direction: 'up'},
-      // {
-      //   type: "textMessage", 
-      //   text: "Hey! Before you head out into town I should explain some things to you."
-      // },
-      // {
-      //   type: "textMessage", 
-      //   text: "To explore the town, use the 'W', 'A', 'S', and 'D' keys on your keyboard. If you would like to chat to any of the villagers, press 'ENTER' to spark up a conversation. Good luck!"
-      // },
-      // {who: 'villager', type: 'walk', direction: 'left'},
-      // {who: 'villager', type: 'walk', direction: 'up'},
-      // {who: 'villager', type: 'walk', direction: 'up'},
-      // {who: 'villager', type: 'walk', direction: 'left'},
-      // {who: 'villager', type: 'walk', direction: 'up'},
+      {who: 'hero', type: 'walk', direction: 'right'},
+      {who: 'hero', type: 'walk', direction: 'right'},
+      {who: 'hero', type: 'walk', direction: 'right'},
+      {who: 'hero', type: 'walk', direction: 'down'},
+      {who: 'hero', type: 'walk', direction: 'down'},
+      {who: 'villager', type: 'walk', direction: 'left'},
+      {who: 'villager', type: 'walk', direction: 'left'},
+      {who: 'villager', type: 'walk', direction: 'up'},
+      {
+        type: "textMessage", 
+        text: "Hey! Before you head out into town I should explain some things to you."
+      },
+      {
+        type: "textMessage", 
+        text: "To explore the town, use the 'W', 'A', 'S', and 'D' keys on your keyboard. If you would like to chat to any of the villagers, press 'ENTER' to spark up a conversation. Good luck!"
+      },
+      {who: 'villager', type: 'walk', direction: 'left'},
+      {who: 'villager', type: 'walk', direction: 'up'},
+      {who: 'villager', type: 'walk', direction: 'up'},
+      {who: 'villager', type: 'walk', direction: 'left'},
+      {who: 'villager', type: 'walk', direction: 'up'},
     ];
 
     this.cutSceneSpaces = {
       ['15, 14']: [
         {
           events: [
-            {who: 'villager', type: "walk", direction: 'right'},
-            {who: 'villager', type: "walk", direction: 'right'},
-            {who: 'villager', type: "stand", direction: 'down', time: 200},
-            {who: 'hero', type: "walk", direction: 'up'},
-            {
-              type: "textMessage", 
-              text: "Good luck out there!",
-            },
-            {who: 'villager', type: "walk", direction: 'left'},
-            {who: 'villager', type: "walk", direction: 'left'},
+            {type: "changeMap", map: "village"}
           ]
         }
       ],
-    }
-  }
-
-  // FIXME: capture the position of the hero whenever they stop walking. 
-  checkForFootstepCutscene() {
-    const hero = this.characters['hero'];
-    const pos = this.scene.area.worldToTileXY(hero.sprite.x, hero.sprite.y)
-    const match = this.cutSceneSpaces[`${pos.x}, ${pos.y}`];
-
-    if (!this.isCutScenePlaying && match) {
-      this.startCutScene(match[0].events)
-    }
-  }
-
-  async startCutScene(events) {
-    this.isCutScenePlaying = true;
-
-    // start a loop of async events
-    for (let i = 0; i < events.length; i++) {
-      const eventHandler = new OverworldEvent({
-        event: events[i],
-        map: this,
-      })
-      await eventHandler.init();
-    }
-    this.isCutScenePlaying = false;
-
-    // reset NPC to do their idle behavior
-    Object.values(this.characters).forEach(c => c.doBehaviorEvent(this))
-  }
-
-  nextPosition(initialX, initialY, direction) {
-    let x = this.scene.area.worldToTileX(initialX);
-    let y = this.scene.area.worldToTileY(initialY);
-    const size = 1;
-    if (direction === 'left') {
-      x -= size;
-    } else if (direction === 'right') {
-      x += size;
-    } else if (direction === 'up') {
-      y -= size;
-    } else if (direction === 'down') {
-      y += size;
-    }
-
-    return {x, y};
-  }
-
-  checkForActionCutscene() {
-    const hero = this.characters['hero'];
-    const nextCoords = this.nextPosition(hero.sprite.x, hero.sprite.y, hero.direction);
-    const match = Object.values(this.characters).find(object => {
-      let characterX = this.scene.area.worldToTileX(object.sprite.x);
-      let characterY = this.scene.area.worldToTileY(object.sprite.y);
-      return `${characterX}, ${characterY}` === `${nextCoords.x}, ${nextCoords.y}`
-    })
-    
-    // TODO: start talking cutscene 
-    if (!this.isCutScenePlaying && match && match.talking.length) {
-      this.startCutScene(match.talking[0].events)
     }
   }
 };
